+++ data/colors_v2/index.ts
@@ -1,7 +1,7 @@
-import light from './light'
-import dark from './dark'
-import dark_dimmed from './dark_dimmed'
-import dark_high_contrast from './dark_high_contrast'
+import light from './themes/light'
+import dark from './themes/dark'
+import dark_dimmed from './themes/dark_dimmed'
+import dark_high_contrast from './themes/dark_high_contrast'
 
 export default {
   light,
