+++ data/colors_v2/vars/deprecated.ts
