+++ data/colors_v2/vars/global_dark.ts
@@ -1,0 +1,92 @@
+import {alpha, get} from '../../../src/utils'
+
+export default {
+  fg: {
+    default: get('scale.gray.1'),
+    muted: get('scale.gray.3'),
+    inactive: get('scale.gray.5'),
+    onEmphasis: get('scale.white')
+  },
+  canvas: {
+    default: get('scale.gray.9'),
+    overlay: get('scale.gray.8'),
+    inset: get('scale.black')
+  },
+  border: {
+    default: get('scale.gray.6'),
+    divider: get('scale.gray.7')
+  },
+  shadow: {
+    small: '0 0 transparent',
+    medium: (theme: any) => `0 3px 6px ${get('scale.black')(theme)}`,
+    large: (theme: any) => `0 8px 24px ${get('scale.black')(theme)}`,
+    extraLarge: (theme: any) => `0 12px 48px ${get('scale.black')(theme)}`
+  },
+
+  // Roles
+  neutral: {
+    fg: get('scale.gray.4'),
+    emphasis: get('scale.gray.5'),
+    highlighter: alpha(get('scale.gray.4'), 0.4),
+    muted: alpha(get('scale.gray.4'), 0.1)
+  },
+  accent: {
+    fg: get('scale.blue.4'),
+    emphasis: get('scale.blue.5'),
+    highlighter: alpha(get('scale.blue.4'), 0.4),
+    muted: alpha(get('scale.blue.4'), 0.1)
+  },
+  success: {
+    fg: get('scale.green.4'),
+    emphasis: get('scale.green.5'),
+    highlighter: alpha(get('scale.green.4'), 0.4),
+    muted: alpha(get('scale.green.4'), 0.1)
+  },
+  warning: {
+    fg: get('scale.yellow.4'),
+    emphasis: get('scale.yellow.5'),
+    highlighter: alpha(get('scale.yellow.4'), 0.4),
+    muted: alpha(get('scale.yellow.4'), 0.1)
+  },
+  severe: {
+    fg: get('scale.orange.4'),
+    emphasis: get('scale.orange.5'),
+    highlighter: alpha(get('scale.orange.4'), 0.4),
+    muted: alpha(get('scale.orange.4'), 0.1)
+  },
+  danger: {
+    fg: get('scale.red.4'),
+    emphasis: get('scale.red.5'),
+    highlighter: alpha(get('scale.red.4'), 0.4),
+    muted: alpha(get('scale.red.4'), 0.1)
+  },
+  done: {
+    fg: get('scale.purple.4'),
+    emphasis: get('scale.purple.5'),
+    highlighter: alpha(get('scale.purple.4'), 0.4),
+    muted: alpha(get('scale.purple.4'), 0.1)
+  },
+  sponsors: {
+    fg: get('scale.pink.4'),
+    emphasis: get('scale.pink.5'),
+    highlighter: alpha(get('scale.pink.4'), 0.4),
+    muted: alpha(get('scale.pink.4'), 0.1)
+  },
+
+  // Only meant for Primer components
+  primer: {
+    canvas: {
+      backdrop: alpha(get('scale.black'), 0.8), // use for modal/dialogs
+      sticky: alpha(get('scale.gray.9'), 0.95) // use for sticky headers
+    },
+    border: {
+      active: '#F78166', // coral
+      contrast: alpha(get('scale.white'), 0.2) // use to increase contrast
+    },
+    shadow: {
+      highlight: '0 0 transparent', // top highlight
+      inset: '0 0 transparent', // top inner shadow
+      focus: (theme: any) => `0 0 0 3px ${get('scale.blue.8')(theme)}` // blue focus ring
+    }
+  }
+}
