+++ data/colors_v2/vars/product_dark.ts
