+++ data/colors_v2/vars/product_light.ts
